api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
mysqlclient==2.2.4
websockets==12.0